Corrigiendo algunas lineas

+++ tp1_RamosAlvaroMaximiliano/src/ar/edu/unju/fi/ejercicio1/Main.java
@@ -3,13 +3,13 @@ package ar.edu.unju.fi.ejercicio1;
 public class Main {
 
 	public static void main(String[] args) {
-		float numUno = 8;
-		float numDos = 5;
-		float numTres = 1;
-		float numCuatro = 20;
-		float numCinco = 9;
+		Integer numUno = 8;
+		Integer numDos = 5;
+		Integer numTres = 1;
+		Integer numCuatro = 20;
+		Integer numCinco = 9;
 		
-		float promedio = (numUno+numDos+numTres+numCuatro+numCinco) / 5;
+		Integer promedio = (numUno+numDos+numTres+numCuatro+numCinco) / 5;
 		System.out.println("El promedio es: " + promedio);
 	}
 }
